# swagger_output_autogen.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "API documentation using Swagger",
    "version": "1.0.0"
  },
  "host": "localhost:4000",
  "servers": [
    {
      "url": "/api"
    }
  ],
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization",
      "description": "JWT Authorization header using the Bearer scheme"
    }
  },
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/products/upload/{id}": {
      "post": {
        "tags": [
          "Product"
        ],
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "files",
            "in": "formData",
            "type": "array",
            "required": true,
            "description": "Some description...",
            "collectionFormat": "multi",
            "items": {
              "type": "file"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/products/upload1/{id}": {
      "post": {
        "tags": [
          "Product"
        ],
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "multFiles",
            "in": "formData",
            "type": "array",
            "required": true,
            "description": "Some description...",
            "collectionFormat": "multi",
            "items": {
              "type": "file"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/posts/upload/{id}": {
      "post": {
        "tags": [
          "Post"
        ],
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "singleFile",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Some description..."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/upload/image": {
      "post": {
        "tags": [
          "Upload"
        ],
        "description": "",
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/upload/images": {
      "post": {
        "tags": [
          "Upload"
        ],
        "description": "",
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
...
}